@@ -23,54 +23,53 @@
     "url": "https://github.com/mikechabot/react-boilerplate/issues"
   },
   "dependencies": {
-    "autoprefixer": "^6.3.6",
-    "axios": "^0.15.2",
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.2.7",
-    "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.0.15",
-    "babel-preset-stage-0": "^6.0.15",
+    "autoprefixer": "^6.7.7",
+    "axios": "^0.16.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^6.4.1",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "chai": "^3.5.0",
-    "coveralls": "^2.11.15",
-    "cross-env": "^3.1.3",
-    "css-loader": "^0.26.0",
+    "coveralls": "^2.13.0",
+    "cross-env": "^4.0.0",
+    "css-loader": "^0.28.0",
     "del": "^2.2.0",
-    "eslint": "^3.10.2",
-    "eslint-config-standard": "^6.2.1",
-    "eslint-loader": "^1.6.1",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-react": "^6.7.1",
-    "eslint-plugin-standard": "^2.0.1",
+    "eslint": "^3.19.0",
+    "eslint-config-standard": "^10.2.0",
+    "eslint-loader": "^1.7.1",
+    "eslint-plugin-promise": "^3.5.0",
+    "eslint-plugin-react": "^6.10.3",
+    "eslint-plugin-standard": "^3.0.1",
     "express": "^4.13.4",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.11.1",
     "html-webpack-plugin": "^2.24.1",
     "istanbul": "^1.0.0-alpha.2",
-    "jsdom": "^9.8.3",
+    "jsdom": "^9.12.0",
     "json-loader": "^0.5.4",
-    "lodash": "^4.16.6",
+    "lodash": "^4.17.4",
     "mocha": "^3.1.2",
-    "moment": "^2.15.2",
-    "node-sass": "^4.1.1",
-    "postcss-loader": "^1.1.0",
-    "react": "^15.4.1",
-    "react-addons-test-utils": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "react-redux": "^5.0.1",
+    "moment": "^2.18.1",
+    "node-sass": "^4.5.2",
+    "postcss-loader": "^1.3.3",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-addons-test-utils": "^15.5.1",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.4",
     "react-router": "^3.0.0",
     "redux": "^3.3.1",
     "redux-entity": "^4.1.1",
-    "redux-logger": "^2.7.4",
-    "redux-mock-store": "^1.0.2",
-    "redux-thunk": "^2.0.1",
-    "sass-loader": "^4.0.2",
-    "sinon": "^1.17.3",
-    "style-loader": "^0.13.0",
+    "redux-logger": "^3.0.1",
+    "redux-mock-store": "^1.2.3",
+    "redux-thunk": "^2.2.0",
+    "sass-loader": "^6.0.3",
+    "sinon": "^2.1.0",
+    "style-loader": "^0.16.1",
     "svg-react-loader": "^0.3.2",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.3",
-    "webpack-dev-server": "^1.12.1"
+    "url-loader": "^0.5.8",
+    "webpack": "^2.3.3",
+    "webpack-dev-server": "^2.4.2"
   }
 }
