@@ -12,6 +12,6 @@ export default function configureStore (initialState) {
     return createStore(
         rootReducer,
         initialState,
-        applyMiddleware(thunk, logger())
+        applyMiddleware(thunk, logger)
     );
 }
