@@ -17,7 +17,7 @@ export default function configureStore (initialState) {
     return createStore(
         rootReducer,
         initialState,
-        // applyMiddleware(thunk),              // USE ME FOR PROD!
-        applyMiddleware(thunk, logger())        // DON'T USE ME FOR PROD!
+        // applyMiddleware(thunk),          // USE ME FOR PROD!
+        applyMiddleware(thunk, logger)      // DON'T USE ME FOR PROD!
     );
 }
